Start work on adding web links to todos
For ease of use one will be able to a shortcode for links. This will be
for two uses:

1. Make it quicker to type a common link (e.g. to a github repo or a
jira ticket).

2. To aid display

As an example, you will be able to write "github!mattyhall/brenda" to
get https://github.com/mattyhall/brenda.

These will be user configurable so we need a new database table and
command to add them.